#1 Convert deletion to a endless loop
This tool should not be run in a container with CRON because doing so may lead
to keeping the container running when the main process exists (open processes
keep containers awake at night).

Instead this commit introduces 2 go routines:
- The first go-routine watches for process signals which may signal that the
  application should be exited.
- The second one runs a ticker that executes the deletion routine every now
  and then (according the given interval though)

Testing go-routines is overly tedious. Thus, this commit removes the CLI tests
drastically.